requirements.txt
# torch 
# torchvision 
opencv-python 
prettytable
jupyter 
tqdm
torchsummary

